dashboard/requirements.txt
streamlit==1.24.0
plotly==5.15.0
pandas==2.0.3
azure-monitor-query==1.2.0
azure-identity==1.13.0
mlflow==2.4.1
python-dotenv==1.0.0
numpy==1.25.1
azure-storage-file-datalake==12.12.0
pandas-profiling==3.2.0
